+++ RimozioneDatiProfili.ps1
@@ -1,10 +1,4 @@
 #------------------------------------------------------------------------
-# Start-Sleep 5
-param(
-  [Parameter(Mandatory = $false)]
-  [boolean]$Anteprima = $false
-)
-$Global:Anteprima = $Anteprima
 #------------------------------------------------------------------------
 $global:excludeUsers = @()
 $global:esclusifileName = 'UtentiEsclusi.txt'
@@ -15,7 +9,7 @@ $global:scriptName = [System.IO.Path]::GetFileNameWithoutExtension($MyInvocation
 $global:logFileExt = "log"
 $global:logFileName = $global:scriptName + "-" + $today + "." + $global:logFileExt
 $global:logFilePath = Join-Path -Path $global:currentPath -ChildPath $global:logFileName
-$global:logExpiration = 1 # giorni
+$global:logExpiration = 7 # giorni
 #------------------------------------------------------------------------
 
 function CaricaEsclusi {
@@ -71,7 +65,7 @@ function Get-UnixTimestamp {
   return $unixTime
 }
 
-function ConvertToHHMMSS {
+function Convert-To-HHMMSS {
   [CmdletBinding()]
   param (
     [Parameter(Mandatory = $true)]
@@ -119,6 +113,7 @@ function RimozioneDatiProfiliRoutine {
     Write-Host $msg
     Write-Host ""
     Write-Log $msg
+    Write-Log "[ERRORE]: $_"
     PremereUnTasto
     Return
   }
@@ -135,16 +130,11 @@ function RimozioneDatiProfiliRoutine {
     PremereUnTasto
     Return
   }
-  if ( $Anteprima ) {
-    Write-Host ""
-    Write-Host "------- MODALITA' ANTEPRIMA, I DATI NON VERRANNO CANCELLATI -------"
-    Write-Host ""
-  }
   
   Write-Host ""
   Write-Host "ATTENZIONE, i dati di ${profileCount} profilo/i utente verranno cancellati definitivamente"
   Write-Host "L'OPERAZIONE NON POTRA' ESSERE ANNULLATA"
-  Write-Host "Il tasto ESC interrompera' il processo di cancellazione alla fine dell'operazione in corso"
+  Write-Host "Premere ESC per interrompere il processo di cancellazione alla fine dell'operazione in corso"
   Write-Host ""
   $risposta = Read-Host "Confermare? (S|[N])"
   if ($risposta -ne "S") {
@@ -167,7 +157,6 @@ function RimozioneDatiProfiliRoutine {
 
     if ([System.Console]::KeyAvailable) {
       $keyInfo = [System.Console]::ReadKey($true)
-  
       # Controlla se l'utente ha premuto il tasto ESC
       if ($keyInfo.Key -eq "Escape") { 
         # Termina il loop quando viene premuto il tasto ESC.
@@ -189,51 +178,82 @@ function RimozioneDatiProfiliRoutine {
     $counter++
     Write-Host
     Write-Host "Eliminazione del profilo $counter $user ($toDo/$profileCount) ${tempoStrascorsoLabel} ${tempoStimatoLabel}"
+    Write-Log "Eliminazione dati utente ${user} ($UserHomePath)"
     $toDo--
-        
+    
     $start = Get-UnixTimestamp
-        
+    
+    # Primo tentaivo di eliminazione
     try {
-      if ( -not $Anteprima ) {
-        $UserProfile | Remove-WmiObject
-        Write-Log "Eliminazione dati utente ${user} ($UserHomePath)"
-      }
+      $UserProfile | Remove-WmiObject
       $elementiCancelati++
     }
     catch {
-      Write-Host
-      Write-Host "Errore: " $_.Exception.Message
-      Write-Log $_.Exception.Message
-      $errori++
-      if ( -not $Anteprima ) {
-        $userDirectoryPath = $UserProfileToDelete.LocalPath
-        if ((Test-Path -Path $userDirectoryPath) -and $userDirectoryPath.ToLower().StartsWith("c:\users\") -and $userDirectoryPath.Length -gt 9 ) {
-          Remove-Item -Path $userDirectoryPath -Recurse -Force -ErrorAction SilentlyContinue
-        }
+      Write-Host "Errore: $_"
+      Write-Log "[ERRORE]: $_"
+    }
+
+    $userDirectoryPath = $UserProfile.LocalPath
+    $RegProfilePath = "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\ProfileList\"
+    $RegProfileKey = $RegProfilePath + $UserProfile.SID
+
+    # Verifica l'effettiva eliminazione della chiave di registro
+    if ((Test-Path -Path $RegProfileKey) -and ($RegProfilePath -ne $RegProfileKey)) {
+      try {
+        $msg = "Eliminazione chiave di registro $RegProfileKey"
+        Write-Host $msg
+        Write-Log $msg
+        Remove-Item -Path $RegProfileKey -Recurse
       }
+      catch {
+        Write-Host "Errore: $_"
+        Write-Log "[ERRORE]: $_"
+      }
+    }
+
+    # Verifica l'effettiva eliminazione della directory
+    if ((Test-Path -Path $userDirectoryPath) -and $userDirectoryPath.ToLower().StartsWith("c:\users\") -and $userDirectoryPath.Length -gt 9 ) {
+      try {
+        $msg = "Eliminazione directory $userDirectoryPath"
+        Write-Host $msg
+        Write-Log $msg
+        Remove-Item -Path $userDirectoryPath -Recurse -Force -ErrorAction SilentlyContinue
+      }
+      catch {
+        Write-Host "Errore: $_" 
+        Write-Log "[ERRORE]: $_"
+      }
+    }
+
+    # Verifica l'effettiva eliminazione della chiave di registro e della directory
+    if ((Test-Path -Path $RegProfileKey) -or (Test-Path -Path $userDirectoryPath)) {
+      $errori++
+      $msg = "Errore imprevisto, impossibile procedere con la cancellazione dei dati dell'utente"
+      Write-Host $msg
+      Write-Log "[ERRORE]: $msg"
     }
   
     $stop = Get-UnixTimestamp
     
     $d = $stop - $start;
     $totalSeconds = $totalSeconds + $d
-  
     $tempoStimato = ($totalSeconds / $counter) * $toDo;
   
-    $tempoStrascorsoHHMMSS = ConvertToHHMMSS -unixTimestamp $totalSeconds
-    $tempoStimatoHHMMSS = ConvertToHHMMSS -unixTimestamp $tempoStimato
+    $tempoStrascorsoHHMMSS = Convert-To-HHMMSS -unixTimestamp $totalSeconds
+    $tempoStimatoHHMMSS = Convert-To-HHMMSS -unixTimestamp $tempoStimato
     $tempoStrascorsoLabel = " Tempo trascorso: $tempoStrascorsoHHMMSS"
     $tempoStimatoLabel = " Tempo stimato: $tempoStimatoHHMMSS"
 
   }
   
-  $tempoStrascorsoHHMMSS = ConvertToHHMMSS -unixTimestamp $totalSeconds
+  $tempoStrascorsoHHMMSS = Convert-To-HHMMSS -unixTimestamp $totalSeconds
   Write-Host ""
   Write-Host "--------------------------------------------------------"
   Write-Host "Fine pulizia profili"
   Write-Host "Profili cancellati $elementiCancelati"
   Write-Host "Tempo trascorso $tempoStrascorsoHHMMSS"
   Write-Host "Errori $errori"
+  Write-Host "--------------------------------------------------------"
   Write-Host ""
   PremereUnTasto
 }