+++ CHANGELOG.md
@@ -4,22 +4,25 @@ Tutti i cambiamenti notevoli al progetto saranno documentati in questo file.
 
 Il formato si basa su [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [Unreleased] - 2023-11-22
+## [0.2.0] - 2023-11-23
 
 ### Aggiunto
 
 - Tasto ESC per interrompere il processo di cancellazione
-- File log, vengono cancellati i logo più vecchi di 15 giorni
+- File log, viene creato un file di log con la data del giorno. I log più vecchi di 7 giorni vengono cancellati
+- LICENSE.txt
 
 ### Cambiato
 
 - Riorganizzazione del codice
 - Modifica messaggi di dialogo utente
 - Gestione della versione con git
+- Controllo cancellazione directory e chiavi di registro
 
 ### Fixed
 
 - Gestione dell'errore che si verifica quando la query Get-WmiObject non viene eseguita.
+- Fix gestione errori
 
 ## [0.1.0] - 2023-11-16